Aulas note

+++ Linguagem de Programação III/Aula03_02/src/business/Quadrado.java
@@ -2,9 +2,8 @@ package business;
 
 public class Quadrado extends Figura {
 
-	private int lado;
-	
-	public void setLado(int lado) { this.lado = lado; }
+	private float lado;
+	public void setLado(float lado) { this.lado = lado; }
 	
 	@Override
 	public float calcularArea() {
